chore: add UV sensor functionality

+++ include/sensors/pressure.h
@@ -1,2 +1,2 @@
 int setup_pressure();
-double read_pressure();
+float read_pressure();
